+++ landingzones/landingzone_caf_foundations/foundations.tf
@@ -1,3 +1,4 @@
+## calling the blueprints
 module "blueprint_foundations_accounting" {
     source                              = "./blueprint_foundations_accounting/"
 
@@ -16,7 +17,7 @@ module "blueprint_foundations_security" {
 
     location                            = var.global_settings.location_map.region1
     tags_hub                            = var.global_settings.tags_hub
-    resource_groups_hub                 = module.blueprint_foundations_accounting.resource_group_hub_names
+    resource_groups_hub                 = module.blueprint_foundations_accounting.resource_group_operations
     log_analytics                       = module.blueprint_foundations_accounting.log_analytics_workspace
 
     security_settings                   = var.security_settings
@@ -28,5 +29,6 @@ module "blueprint_foundations_governance" {
     tags_hub                            = var.global_settings.tags_hub
     location                            = var.global_settings.location_map.region1
     log_analytics                       = module.blueprint_foundations_accounting.log_analytics_workspace
+    
     governance_settings                 = var.governance_settings
 }