+++ landingzones/landingzone_hub_spoke/hub_network/bastion/output.tf
@@ -1,0 +1,4 @@
+output "bastion" {
+  depends_on = [azurerm_bastion_host.azurebastion]
+  value = azurerm_bastion_host.azurebastion
+}