+++ landingzones/landingzone_hub_spoke/variables_landingzone.tf
@@ -1,0 +1,7 @@
+variable "core_networking" {
+    description = "(Required) Ccomplex configuration object that describes the hub network configuration"
+}
+variable "rg_network" {
+    description = "(Required) Name of the resource groups for the hub network"
+}
+