+++ landingzones/landingzone_secure_vnet_dmz/net_core/variables.tf
@@ -72,3 +72,5 @@ variable "location" {
 }
 
 variable "rg_network" {}
+
+variable "logged_user_objectId" {}