+++ landingzones/landingzone_secure_vnet_dmz/net_core/vpn_gateway/access_policies.tf
@@ -16,3 +16,18 @@ resource "azurerm_key_vault_access_policy" "akv_policy1" {
   ]
 }
 
+resource "azurerm_key_vault_access_policy" "vpn_akv_current_user" {
+  key_vault_id = var.keyvaultid
+
+  tenant_id = data.azurerm_client_config.current.tenant_id
+  object_id = var.logged_user_objectId 
+
+  key_permissions = []
+
+  secret_permissions = [
+    "set",
+    "get",
+    "list",
+    "delete",
+  ]
+}