+++ landingzones/landingzone_starter/main.tf
@@ -1,5 +1,6 @@
 provider "azurerm" {
-  version = "<=1.44"
+  version = "~>2.4.0"
+  features {}
 }
 
 terraform {