+++ landingzones/landingzone_vdc_demo/blueprint_networking_shared_egress/main.tf
@@ -4,10 +4,6 @@ terraform {
     }
 }
 
-provider "azurerm" {
-  version = "<= 1.44"
-}
-
 data "azurerm_subscription" "current" {
 }