+++ landingzones/landingzone_vdc_demo/bp_operations.tf
@@ -6,9 +6,9 @@
     caf_foundations_accounting         = local.caf_foundations_accounting
     global_settings                    = local.global_settings
 
-    resource_groups_operations         = lookup(local.caf_foundations_accounting.resource_group_hub_names, "HUB-OPERATIONS", null)
-    asr_rg                             = lookup(local.caf_foundations_accounting.resource_group_hub_names, "HUB-OPERATIONS", null)  
-    auto_rg                            = lookup(local.caf_foundations_accounting.resource_group_hub_names, "HUB-OPERATIONS", null)
+    resource_groups_operations         = lookup(local.caf_foundations_accounting.resource_group_operations, "name", null)
+    asr_rg                             = lookup(local.caf_foundations_accounting.resource_group_operations, "name", null)  
+    auto_rg                            = lookup(local.caf_foundations_accounting.resource_group_operations, "name", null)
     
     asr_config                         = var.asr_config
     auto_config                        = var.auto_config