+++ CHANGELOG.md
@@ -1,3 +1,20 @@
+## v3.0.2003 (March 2020)
+
+FEATURES:
+
+* **rover :** upgrade to rover 2004.0211 - added support for Terraform v0.12.24
+* **added support for azurecaf provider :** Overall usage of new [azurecaf provider](https://github.com/aztfmod/terraform-provider-azurecaf) for naming convention
+* **landingzone_hub_spoke :** introducing landingzone_hub_spoke as a sample of hub and spoke topology [#13](
+https://github.com/Azure/caf-terraform-landingzones/issues/13)
+* **landingzone_caf_foundations :** Added toggle feature for management group deployment control (disabled by default)
+* **landingzone_caf_foundations :** Added support for azurerm 2.4 [#23](
+https://github.com/Azure/caf-terraform-landingzones/issues/23)
+* **landingzone_vdc_demo:** Added support for azurerm 2.4 [#23](
+https://github.com/Azure/caf-terraform-landingzones/issues/23)
+* **landingzone_secure_vnet_dmz:** Added support for azurerm 2.4 [#23](
+https://github.com/Azure/caf-terraform-landingzones/issues/23)
+
+
 ## v2.2.2002 (February 2020)
 
 FEATURES: